Lab3/forum/requirements.txt
asgiref==3.5.2
Django==4.0.5
django-etc==1.3.1
django-hitcount==1.3.5
django-resized==1.0.0
django-taggit==3.0.0
django-tinymce==3.4.0
sqlparse==0.4.2
